requirements.txt
cheroot == 8.4.5          # Used only in deprecated code and will be removed in a later release.
portend == 2.5
djangorestframework == 3.15.2
djangorestframework-bulk == 0.2.1
djangorestframework-filters == 0.10.2
django-bootstrap3 == 15.0.0
django-filter == 2.4.0
html5lib == 1.0.1 ; python_version < '3.10'
html5lib == 1.1 ; python_version >= '3.10'
jdcal == 1.4.1
Markdown == 3.1.1
openpyxl == 3.0.9
lxml == 5.1.0
PyJWT == 2.4.0
urllib3 < 2
requests == 2.32.3
python-dateutil == 2.8.1
pillow >= 8.4.0
psutil == 5.9.6
psycopg2-binary == 2.8.6  ; python_version < '3.10'
psycopg2-binary == 2.9.9  ; python_version >= '3.10'
# Python packaging is broken on ubuntu 20 for higher versions of setuptools
setuptools == 46.1 ; python_version < '3.10'
setuptools ; python_version >= '3.11'
setuptools-rust == 0.12.1
pysftp == 0.2.9
channels == 4.0.0
attrs == 23.1.0
daphne == 4.0.0
Twisted[tls,http2]
https://github.com/frePPLe/django/tarball/frepple_8.3

pymssql